Moved chart to common component
