+++ MyMoney.Web/ClientApp/src/app/pages/index.ts
@@ -10,6 +10,5 @@ export * from './budgets/edit/edit.budgets.component';
 export * from './incomes/incomes.component';
 export * from './incomes/add/add.incomes.component';
 export * from './incomes/edit/edit.incomes.component';
-export * from './home/chart/chart.component';
 export * from './transactions/import/import.transactions.component';
 export * from './incomes/import/import.incomes.component';
