+++ MyMoney.Web/ClientApp/src/app/shared/components/chart/chart-data-provider.interface.ts
