+++ MyMoney.Web/ClientApp/src/app/shared/components/chart/chart.component.ts
