+++ MyMoney.Web/ClientApp/src/app/shared/components/chart/index.ts
@@ -1,0 +1,2 @@
+export * from './chart.component';
+export * from './chart-data-provider.interface';
